Added RESTRICTED_CAST_FROM_ASCII define

+++ src/xpiks-qt/AutoComplete/autocompleteworker.cpp
@@ -97,7 +97,7 @@ namespace AutoComplete {
         QStringList completionsList;
         QSet<QString> completionsSet;
 
-        int size = completions.size(), i = 0;
+        int size = (int)completions.size(), i = 0;
         completions.reserve(size);
         completionsSet.reserve(size);
